refactor: move test utilities to internal package and enhance logging functionality
