+++ internal/testutil/encoders_test.go
@@ -3,9 +3,8 @@ package testutil_test
 import (
 	"testing"
 
+	"go.robertomontagna.dev/zapfluent/internal/testutil"
 	"go.uber.org/zap/zapcore"
-
-	"go.robertomontagna.dev/zapfluent/testutil"
 
 	. "github.com/onsi/gomega"
 )
