+++ pkg/core/error_handler_test.go
@@ -4,12 +4,11 @@ import (
 	"errors"
 	"testing"
 
+	"go.robertomontagna.dev/zapfluent/internal/testutil/stubs"
 	"go.uber.org/zap/zapcore"
 
-	"go.robertomontagna.dev/zapfluent/pkg/core"
-	"go.robertomontagna.dev/zapfluent/testutil/stubs"
-
 	. "github.com/onsi/gomega"
+	"go.robertomontagna.dev/zapfluent/pkg/core"
 )
 
 var (
